working with fuctions
