+++ mohammed/function.c
@@ -1,0 +1,11 @@
+#include <stdio.h>
+#include <stdlib.h>
+
+int main()
+{
+
+
+
+	return 0;
+
+}
